Extract a common project for GitHubApp authentication.
