+++ GitHub.Collectors.sln
@@ -18,6 +18,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core.Collectors", "CEDAR.Co
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core.Collectors.Tests", "CEDAR.Core.Collector\Core.Collectors.Tests\Core.Collectors.Tests.csproj", "{AC2CC7AC-7CC7-435F-9EC5-6AB12098B7DC}"
 EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GitHub.Common", "GitHub.Common\GitHub.Common.csproj", "{02E64C94-3EEB-4D18-930A-8D6B6CD13539}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -88,6 +90,18 @@ Global
 		{AC2CC7AC-7CC7-435F-9EC5-6AB12098B7DC}.Release|x64.Build.0 = Release|Any CPU
 		{AC2CC7AC-7CC7-435F-9EC5-6AB12098B7DC}.Release|x86.ActiveCfg = Release|Any CPU
 		{AC2CC7AC-7CC7-435F-9EC5-6AB12098B7DC}.Release|x86.Build.0 = Release|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Debug|x64.Build.0 = Debug|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Debug|x86.Build.0 = Debug|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Release|Any CPU.Build.0 = Release|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Release|x64.ActiveCfg = Release|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Release|x64.Build.0 = Release|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Release|x86.ActiveCfg = Release|Any CPU
+		{02E64C94-3EEB-4D18-930A-8D6B6CD13539}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
