+++ GitHub.Common/Authentication/GitHubAppAuthentication.cs
@@ -15,10 +15,10 @@ using System.Net.Http;
 using Microsoft.CloudMine.Core.Collectors.Web;
 using Newtonsoft.Json.Linq;
 using Microsoft.CloudMine.Core.Collectors.Authentication;
-using Microsoft.CloudMine.GitHub.Collectors.Web;
+using Microsoft.CloudMine.GitHub.Web;
 using System.Collections.Concurrent;
 
-namespace Microsoft.CloudMine.GitHub.Collectors.Authentication
+namespace Microsoft.CloudMine.GitHub.Authentication
 {
     public class GitHubAppAuthentication : IAuthentication
     {
@@ -28,7 +28,7 @@ namespace Microsoft.CloudMine.GitHub.Collectors.Authentication
         private const int JwtExpiry = 60 * 8; // 8 mins
 
         private string organization;
-        private GitHubHttpClient httpClient;
+        private GitHubHttpClientBase httpClient;
         private string apiDomain;
         private readonly int appId;
         private readonly string gitHubAppKeyVaultUri;
@@ -41,7 +41,7 @@ namespace Microsoft.CloudMine.GitHub.Collectors.Authentication
         private static ConcurrentDictionary<string, Tuple<DateTime, string>> TokenCache = new ConcurrentDictionary<string, Tuple<DateTime, string>>();
         private static ConcurrentDictionary<string, string> OrgNameToInstallationIdMap = new ConcurrentDictionary<string, string>();
 
-        public GitHubAppAuthentication(int appId, GitHubHttpClient httpClient, string organization, string apiDomain, string gitHubAppKeyVaultUri, bool useInteractiveLogin)
+        public GitHubAppAuthentication(int appId, GitHubHttpClientBase httpClient, string organization, string apiDomain, string gitHubAppKeyVaultUri, bool useInteractiveLogin)
         {
             this.appId = appId;
             this.gitHubAppKeyVaultUri = gitHubAppKeyVaultUri;
