+++ GitHub.Common/Web/GitHubHttpClientBase.cs
@@ -1,0 +1,302 @@
+﻿// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Net;
+using System.Net.Http;
+using System.Net.Http.Headers;
+using System.Threading.Tasks;
+using Microsoft.CloudMine.Core.Collectors.Authentication;
+using Microsoft.CloudMine.Core.Collectors.Collector;
+using Microsoft.CloudMine.Core.Collectors.Error;
+using Microsoft.CloudMine.Core.Collectors.Telemetry;
+using Microsoft.CloudMine.Core.Collectors.Web;
+using Newtonsoft.Json;
+using Newtonsoft.Json.Linq;
+
+namespace Microsoft.CloudMine.GitHub.Web
+{
+    // ToDo: kivancm: further abstract the common parts of AzureDevOpsHttpClient and GitHubHttpClient into the Core library.
+    public class GitHubHttpClientBase : IWebRequestStatsTracker
+    {
+        protected readonly IHttpClient httpClient;
+        private readonly IRateLimiter rateLimiter;
+        private readonly ITelemetryClient telemetryClient;
+
+        public static ProductInfoHeaderValue GitHubProductInfoHeaderValue = new ProductInfoHeaderValue("CloudMineGitHubCollector", "1.0.0");
+
+        public const int MaxRequestExceptionCount = 10;
+        public static readonly TimeSpan[] DelayBeforeRequestExceptions = Enumerable.Repeat(TimeSpan.FromSeconds(5), MaxRequestExceptionCount).ToArray();
+
+        public int SuccessfulRequestCount => this.successfulRequestCount;
+        private volatile int successfulRequestCount;
+        public int FailedRequestCount => this.failedRequestCount;
+        private volatile int failedRequestCount;
+
+        private readonly RetryRule[] retryRuleCollection;
+
+        public GitHubHttpClientBase(IHttpClient httpClient, IRateLimiter rateLimiter, ITelemetryClient telemetryClient)
+        {
+            this.httpClient = httpClient;
+            this.rateLimiter = rateLimiter;
+            this.telemetryClient = telemetryClient;
+
+            this.successfulRequestCount = 0;
+            this.failedRequestCount = 0;
+
+            this.retryRuleCollection = new RetryRule[]
+            {
+                RetryRules.GatewayTimeoutRetryRule(),
+                RetryRules.BadGatewayRetryRule(),
+                RetryRules.InternalServerErrorRetryRule(),
+                RetryRules.RateLimiterAbuseRetryRule(),
+            };
+        }
+
+        public Task<HttpResponseMessage> GetAsync(string requestUrl, IAuthentication authentication, string apiName, List<HttpResponseSignature> allowlistedResponses)
+        {
+            return this.MakeRequestAsync(requestUrl, authentication, apiName, eTag: string.Empty, allowlistedResponses, () => this.httpClient.GetAsync(requestUrl, authentication, GitHubProductInfoHeaderValue, eTag: string.Empty));
+        }
+
+        public async Task<HttpResponseMessage> MakeRequestAsync(string requestUrl, IAuthentication authentication, string apiName, string eTag, List<HttpResponseSignature> allowlistedResponses, Func<Task<HttpResponseMessage>> httpMethodCallback)
+        {
+            // New request, clear retry rules (attempt indices)
+            foreach (RetryRule retryRule in this.retryRuleCollection)
+            {
+                retryRule.Clear();
+            }
+
+            HttpResponseMessage response = null;
+            int exceptionCount = 0;
+            TimeSpan delayBeforeRetry = TimeSpan.Zero;
+            bool shallRetry = true;
+            while (shallRetry)
+            {
+                if (delayBeforeRetry != TimeSpan.Zero)
+                {
+                    await Task.Delay(delayBeforeRetry).ConfigureAwait(false);
+                }
+
+                await this.rateLimiter.WaitIfNeededAsync(authentication).ConfigureAwait(false);
+
+                DateTime webRequestDateTime = DateTime.UtcNow;
+
+                try
+                {
+                    response = await httpMethodCallback();
+                    (shallRetry, delayBeforeRetry) = await this.ShallRetryAsync(response, requestUrl, authentication.Identity).ConfigureAwait(false);
+                }
+                catch (Exception requestException)
+                {
+                    exceptionCount++;
+                    (shallRetry, delayBeforeRetry) = await this.ShallRetryAsync(requestException, requestUrl, exceptionCount, authentication.Identity).ConfigureAwait(false);
+                    if (!shallRetry)
+                    {
+                        throw requestException;
+                    }
+                }
+                TimeSpan elapsed = DateTime.UtcNow - webRequestDateTime;
+
+                if (response != null)
+                {
+                    this.telemetryClient.TrackRequest(authentication.Identity, apiName, requestUrl, eTag, elapsed, response);
+
+                    await this.rateLimiter.UpdateStatsAsync(authentication.Identity, requestUrl, response).ConfigureAwait(false);
+                }
+            }
+
+            await this.ThrowOnFatalResponseAsync(response, requestUrl, exceptionCount, authentication.Identity, allowlistedResponses).ConfigureAwait(false);
+            this.successfulRequestCount++;
+            return response;
+        }
+
+        private async Task<Tuple<bool, TimeSpan>> ShallRetryAsync(Exception requestException, string requestUrl, int exceptionCount, string identity)
+        {
+            if (exceptionCount >= DelayBeforeRequestExceptions.Length)
+            {
+                // Out of retries.
+                return Tuple.Create(false, TimeSpan.Zero);
+            }
+
+            bool shallRetry = false;
+            TimeSpan delayBeforeRetry = TimeSpan.Zero;
+
+            if (requestException is AggregateException aggregateException)
+            {
+                foreach (Exception innerException in aggregateException.InnerExceptions)
+                {
+                    (shallRetry, delayBeforeRetry) = await this.ShallRetryAsync(innerException, requestUrl, exceptionCount, identity);
+                    if (shallRetry)
+                    {
+                        // Found a signature match for one of the inner exceptions, retry.
+                        return Tuple.Create(shallRetry, delayBeforeRetry);
+                    }
+                }
+
+                // No match for inner exceptions, return false.
+                return Tuple.Create(false, TimeSpan.Zero);
+            }
+
+            if (requestException is HttpRequestException && requestException.Message.Equals("Error while copying content to a stream."))
+            {
+                // Known flaky server issue with GitHub.
+                shallRetry = true;
+                delayBeforeRetry = DelayBeforeRequestExceptions[exceptionCount - 1];
+            }
+
+            if (shallRetry)
+            {
+                // Track web exception in telemetry. No need to track fatal (non-retried) exceptions since they will be tracked elsewhere.
+
+                Dictionary<string, string> properties = new Dictionary<string, string>()
+                {
+                    { "Url", requestUrl },
+                    { "Identity", identity },
+                    { "AttemptIndex", exceptionCount.ToString() },
+                    { "Retried", true.ToString() },
+                    { "Fatal", false.ToString() },
+                    { "DelayBeforeRetry", delayBeforeRetry.ToString() },
+                };
+                this.telemetryClient.TrackException(requestException, "Web request failed.", properties);
+            }
+
+            return Tuple.Create(shallRetry, delayBeforeRetry);
+        }
+
+        private async Task<Tuple<bool, TimeSpan>> ShallRetryAsync(HttpResponseMessage response, string requestUrl, string identity)
+        {
+            bool shallRetry = false;
+            TimeSpan delayBeforeRetry = TimeSpan.Zero;
+
+            long attemptIndex = -1;
+            foreach (RetryRule retryRule in this.retryRuleCollection)
+            {
+                TimeSpan[] delayBeforeRetries = retryRule.DelayBeforeRetries;
+                if (retryRule.AttemptIndex == delayBeforeRetries.Length)
+                {
+                    // No more attempts for this retry rule, bail out.
+                    continue;
+                }
+
+                bool matches = await retryRule.ShallRetryAsync(response).ConfigureAwait(false);
+                if (matches)
+                {
+                    delayBeforeRetry = delayBeforeRetries[retryRule.AttemptIndex];
+                    shallRetry = true;
+
+                    retryRule.Consume();
+                    attemptIndex = retryRule.AttemptIndex;
+
+                    // If there is a RetryAfter already provided as part of the response, honor that instead of our internal delay.
+                    long retryAfter = RateLimiter.GetRetryAfter(response.Headers);
+                    if (retryAfter != long.MinValue)
+                    {
+                        delayBeforeRetry = TimeSpan.FromSeconds(retryAfter);
+                        await this.rateLimiter.UpdateRetryAfterAsync(identity, requestUrl, response).ConfigureAwait(false);
+                    }
+
+                    break;
+                }
+            }
+
+            if (!response.IsSuccessStatusCode && response.StatusCode != HttpStatusCode.NotModified && shallRetry)
+            {
+                // Track web exception in telemetry. No need to track fatal (non-retried) exceptions since they will be tracked elsewhere.
+                string responseContent = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+                HttpStatusCode responseStatusCode = response.StatusCode;
+
+                Dictionary<string, string> properties = new Dictionary<string, string>()
+                {
+                    { "Url", requestUrl },
+                    { "Identity", identity },
+                    { "ResponseContent", responseContent },
+                    { "ResponseStatusCode", responseStatusCode.ToString() },
+                    { "AttemptIndex", attemptIndex.ToString() },
+                    { "Retried", true.ToString() },
+                    { "Fatal", false.ToString() },
+                    { "DelayBeforeRetry", delayBeforeRetry.ToString() },
+                };
+                Exception exception = new Exception($"Request to url '{requestUrl}' failed with status code: '{responseStatusCode}'. Response content: '{responseContent}'.");
+                this.telemetryClient.TrackException(exception, "Web request failed.", properties);
+            }
+
+            return Tuple.Create(shallRetry, delayBeforeRetry);
+        }
+
+        private async Task ThrowOnFatalResponseAsync(HttpResponseMessage response, string requestUrl, int attemptIndex, string identity, List<HttpResponseSignature> allowlistedResponses)
+        {
+            if (response.IsSuccessStatusCode || response.StatusCode == HttpStatusCode.NotModified)
+            {
+                return;
+            }
+
+            string responseContent = await response.Content.ReadAsStringAsync().ConfigureAwait(false);
+            HttpStatusCode responseStatusCode = response.StatusCode;
+            try
+            {
+                JObject responseContentObject = JObject.Parse(responseContent);
+
+                foreach (HttpResponseSignature allowlistedResponse in allowlistedResponses)
+                {
+                    if (allowlistedResponse.Matches(responseStatusCode, responseContentObject))
+                    {
+                        Dictionary<string, string> allowlistedResponseProperties = new Dictionary<string, string>()
+                        {
+                            { "RequestUrl", requestUrl },
+                            { "ResponseStatusCode", responseStatusCode.ToString() },
+                            { "ResponseContent", responseContentObject.ToString(Formatting.None) },
+                        };
+                        this.telemetryClient.TrackEvent("AllowlistedResponse", allowlistedResponseProperties);
+
+                        return;
+                    }
+                }
+            }
+            catch (Exception)
+            {
+                // Done as best effort, ignore since the response content is already logged.
+            }
+
+            Exception fatalException = new FatalException($"Request to url '{requestUrl}' failed with status code: '{responseStatusCode}'. Response content: '{responseContent}'.");
+            Dictionary<string, string> properties = new Dictionary<string, string>()
+            {
+                { "Url", requestUrl },
+                { "Identity", identity },
+                { "ResponseContent", responseContent },
+                { "ResponseStatusCode", responseStatusCode.ToString() },
+                { "AttemptIndex", attemptIndex.ToString() },
+                { "Retried", false.ToString() },
+                { "Fatal", true.ToString() },
+            };
+            this.telemetryClient.TrackException(fatalException, "Web request failed.", properties);
+            this.failedRequestCount++;
+            throw fatalException;
+        }
+
+        public async Task<JObject> GetAndParseAsJObjectAsync(string requestUrl, IAuthentication authentication, string apiName, List<HttpResponseSignature> allowlistedResponses)
+        {
+            HttpResponseMessage response = await this.GetAsync(requestUrl, authentication, apiName, allowlistedResponses).ConfigureAwait(false);
+            return await HttpUtility.ParseAsJObjectAsync(response).ConfigureAwait(false);
+        }
+
+        public async Task<JArray> GetAndParseAsJArrayAsync(string requestUrl, IAuthentication authentication, string apiName, List<HttpResponseSignature> allowlistedResponses)
+        {
+            HttpResponseMessage response = await this.GetAsync(requestUrl, authentication, apiName, allowlistedResponses).ConfigureAwait(false);
+            return await HttpUtility.ParseAsJArrayAsync(response).ConfigureAwait(false);
+        }
+
+        public async Task<JObject> PostAndParseAsJObjectAsync(string requestUrl, string requestBody, IAuthentication authentication, string apiName, List<HttpResponseSignature> allowlistedResponses)
+        {
+            HttpResponseMessage response = await this.MakeRequestAsync(requestUrl,
+                                                                       authentication,
+                                                                       apiName,
+                                                                       eTag: String.Empty,
+                                                                       allowlistedResponses,
+                                                                       () => this.httpClient.PostAsync(requestUrl, requestBody, authentication, GitHubProductInfoHeaderValue)
+            ).ConfigureAwait(false);
+            return await HttpUtility.ParseAsJObjectAsync(response).ConfigureAwait(false);
+        }
+    }
+}
